requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.9b0
bleach==6.0.0
boto3==1.28.49
botocore==1.31.49
brotlipy==0.7.0
cachetools==5.3.1
cardinality==0.1.1
certifi==2022.12.7
click==8.1.7
cloudpickle==2.2.1
cmake==3.25.0
colored==2.2.3
coloredlogs==15.0.1
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
datasets==2.14.5
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
google-auth==2.23.3
google-auth-oauthlib==1.0.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.59.0
horovod==0.28.1
huggingface-hub==0.16.4
humanfriendly==10.0
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
isort==5.9.3
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
lightning-utilities==0.9.0
lit==15.0.7
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.1
mkl-service==2.4.0
mlprodict==0.9.1883
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==0.4.3
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
netron==7.1.6
networkx==3.0
neural-compressor==2.3.1
notebook==7.0.3
notebook_shim==0.2.3
nptyping==1.4.2
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
onnx2torch==1.5.11
onnxconverter-common==1.14.0
onnxruntime-gpu==1.16.0
onnxsim==0.4.33
opencv-python==4.8.0.76
opencv-python-headless==4.5.4.60
openpyxl==3.1.2
optimum==1.13.2
overrides==3.1.0
packaging==23.2
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
playwright==1.38.0
pluggy==1.3.0
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.7
pyDeprecate==0.3.1
pyee==9.0.4
Pygments==2.16.1
pyparsing==3.0.9
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.5.10
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
ruff==0.1.3
s3transfer==0.6.2
safetensors==0.4.0
schema==0.7.5
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
skl2onnx==1.15.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.11.1
synr==0.6.0
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
terminado==0.17.1
threadpoolctl==3.2.0
timm==0.9.7
tinycss2==1.2.1
tokenize-rt==5.2.0
tokenizers==0.14.1
tomli==1.2.3
torch==1.13.0
torchaudio==0.13.0
torchmetrics==1.2.0
torchview==0.2.6
torchvision==0.14.0
torchviz==0.0.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.34.0
triton==2.0.0
typing_extensions==4.4.0
typish==1.9.3
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.0
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
zstandard==0.21.0
